Fix profile screen
