+++ app/build.gradle
@@ -169,6 +169,7 @@ dependencies {
     implementation "androidx.compose.animation:animation:$compose_version"
     implementation "androidx.compose.ui:ui-tooling-preview:$compose_version"
     implementation 'androidx.constraintlayout:constraintlayout-compose:1.0.1'
+    implementation "androidx.compose.foundation:foundation:$compose_version"
     implementation 'androidx.hilt:hilt-navigation-compose:1.0.0'
     implementation 'com.airbnb.android:lottie-compose:6.0.0'
     implementation 'androidx.activity:activity-compose:1.7.1'
