+++ app/src/androidTest/java/org/listenbrainz/android/LoginActivityTest.kt
@@ -16,7 +16,7 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
-import org.listenbrainz.android.ui.screens.login.LoginActivity
+import org.listenbrainz.android.ui.screens.profile.LoginActivity
 
 @RunWith(AndroidJUnit4ClassRunner::class)
 @LargeTest
