+++ app/src/main/java/org/listenbrainz/android/viewmodel/DashBoardViewModel.kt
@@ -97,5 +97,4 @@ class DashBoardViewModel @Inject constructor(
             }
         }
     }
-    
 }
